(hopefully) make jobs fail when WCSim isn't run, save the geofile.txt references

+++ MakeReference.py
@@ -1,5 +1,6 @@
 #! /bin/env python
 
+import sys
 import glob
 import os
 import subprocess
@@ -32,23 +33,38 @@ def MakeReference(job_num):
     f = open(logfile, 'w')
     subprocess.run(['WCSim', mac, f'{validation_dir}/Generate/macReference/tune/tuning_parameters.mac'], stderr=subprocess.STDOUT, stdout=f, text=True)
     f.close()
-    #Get the relevant histograms from WCSim
+    #Check it worked
     rootfile = mac.rsplit('/',1)[-1].replace('_seed20230628.mac', '.root')
+    if not os.path.isfile(rootfile):
+        print(f'Cannot find expected WCSim output root file: {rootfile}')
+        return False
+    #Get the relevant histograms from WCSim
     subprocess.run([f'{validation_dir}/Generate/daq_readfilemain', rootfile, '0'])
     #Move the histogram file to the reference location
+    found_a_histfile = False
     for branch in ['wcsimrootevent', 'wcsimrootevent2', 'wcsimrootevent_OD']:
         histfile = rootfile.replace('.root', f'_analysed_{branch}.root')
         if os.path.isfile(histfile):
             shutil.move(histfile, f'{validation_dir}/Compare/Reference/{histfile}')
-    #Write the bad printouts information to the reference location
-    badfile = rootfile.replace('.root', '_bad.txt')
+            found_a_histfile = True
     #Get the stats of bad printouts
+    # including writing the bad printouts information to the reference location
+    badfile = rootfile.replace('.root', '_bad.txt')
     bads = {}
     with open(f'{validation_dir}/Compare/Reference/{badfile}', 'w') as f:
         for bad in ['GeomNav1002', '"Optical photon is killed because of missing refractive index"']:
             bads[bad] = int(subprocess.run(['grep', '-c', bad, logfile], stdout=subprocess.PIPE, text=True).stdout)
-            f.write('{bad} {bads[bad]}\n')
+            f.write(f'{bad} {bads[bad]}\n')
     pprint(bads)
+    #Get the geometry text file
+    geofiles = glob.glob('geofile_*.txt')
+    if len(geofiles) == 1:
+        geofile_to_move = geofiles[0]
+    elif len(geofiles) == 2:
+        geofile_to_move = [x for x in geofiles if 'SuperK' not in x][0]
+    shutil.move(geofile_to_move, f'{validation_dir}/Compare/Reference/{geofile_to_move}')
+    #return
+    return not found_a_histfile
 
 def PushToGit(branch_name='new_ref'):
     os.chdir(validation_dir)
@@ -68,10 +84,14 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     jobs_to_run = set(args.job_number)
+    failure = False
     for job in jobs_to_run:
         print(job)
         if job == 0:
             GetSHA()
-        MakeReference(job)
+        failure = failure or MakeReference(job)
     #now all jobs have run, commit to the repo
     PushToGit()
+
+    if failure:
+        sys.exit(-1)
